+++ fe/fe-core/src/main/java/com/starrocks/scheduler/TaskBuilder.java
@@ -1,5 +1,5 @@
 // This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Limited.
-package com.starrocks.job.task;
+package com.starrocks.scheduler;
 
 import com.starrocks.common.util.DebugUtil;
 import com.starrocks.qe.ConnectContext;
