+++ fe/fe-core/src/main/java/com/starrocks/scheduler/TaskManager.java
@@ -1,6 +1,6 @@
 // This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Limited.
 
-package com.starrocks.job.task;
+package com.starrocks.scheduler;
 
 
 import com.clearspring.analytics.util.Lists;
